feat: Form input types
